requirements.txt
scrapy
selenium
pytz
requests
randomcolor
numpy
pandas
python-dateutil